Crate cache_size

A library to quickly get the size and line size of your CPU caches.

Currently this crate only supports x86 CPUs, since it relies on the CPUID instruction, via the raw_cpuid crate. It is a goal to support other architectures; PRs are welcome!

Note that the library will still compile and work on non x86 architectures, but the result of all the cache queries will be None.

Check the Intel 64 and IA-32 Architectures Software Developers Manual for more information on the CPUID instruction.

Enums

CacheType

Functions

cache_line_size
Returns the line size in bytes of level cache with type cache_type.

cache_size
Returns the total size in bytes of level cache with type cache_type.

l1_cache_size
Returns the total size in bytes of the L1 data cache.

l1_cache_line_size
Returns the line size in bytes of the L1 data cache.

l2_cache_size
Returns the total size in bytes of the unified L2 cache.

l2_cache_line_size
Returns the line size in bytes of the unified L2 cache.

l3_cache_size
Returns the total size in bytes of the unified L3 cache.

l3_cache_line_size
Returns the line size in bytes of the unified L3 cache.